ai-server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==11.0.0
numpy>=1.26.0
opencv-python==4.10.0.84